reimpl(swrUI/swrSprite): widget setter family (sprite slots, checked state, values, colors)
24 reverse-hook bodies covering the position/color/value/state setter surface:

swrUI.c: AddSprite, SetSpriteFlag, SetSpriteRect, SetSpriteOffset,
SetSpriteColor helpers (SetListHighlightColor, ApplyListColors,
ApplyFocusColor), NewSpriteElement, SetChecked, ToggleChecked,
ClearGroupChecked, GetNumberValue, GetValue (un-HANGed), SetValueText,
SetValue, GetValueText, SetSlotValue, GetSlotValue, SetUI4,
SetCaretActive, RandomizeSpriteAlpha.

swrSprite.c: FindFreeId, SetPosF, GetUIScale. SetPosF's prototype was
wrong (float args): the binary MOVSXes shorts and converts real px into
a 320x240 design space. Both SetPosF and GetUIScale are forward-hooked
by the res-independence deltas in renderer_hook.cpp, so they join the
GenerateHooks blacklist like stdConsole_GetCursorPos.

Supporting: swrUI_SPRITE_SLOT_FLAG enum (bits decoded from
RenderElementSprites), swrUI_CLASS enum (verified ctor class ids),
swrUI_MSG (partial), swrUI_FLAG additions (STATIC/HOVERED/CHECKED/
radio-group bits); types.h names swrUI_unk2.texture_w/h and the
value_text/value/first_visible fields; data_symbols.syms adds
swrUI_caretEnabled (0x4d87bc).

+++ src/Swr/swrSprite.h
@@ -155,7 +155,7 @@ void swrSprite_SetColor(short id, uint8_t r, uint8_t g, uint8_t b, uint8_t a);
 void swrSprite_SetFlag(short id, unsigned int flag);
 void swrSprite_UnsetFlag(short id, unsigned int flag);
 
-void swrSprite_SetPosF(short id, float x, float y);
+void swrSprite_SetPosF(short id, short x, short y); // converts real screen px -> 320x240 design space, then swrSprite_SetPos
 
 int16_t swrSprite_setCurrentTextPos(int16_t, int16_t);
 short swrSprite_getCurrentTextPos(int16_t*, int16_t*);
